+++ src/action/transfer/send/order.ts
@@ -6,8 +6,6 @@ import {
 } from '../../../action';
 import * as OrderFactory from '../../../order';
 import { IGood, IOwnershipInfo } from '../../../ownershipInfo';
-import { IAttributes as IMoneyTransferActionAttributes } from '../../interact/confirm/moneyTransfer';
-import { IAttributes as IRegisterServiceAttributes } from '../../interact/confirm/registerService';
 import * as SendActionFactory from '../send';
 import { IAttributes as ISendEmailMessageActionAttributes } from './message/email';
 
@@ -27,14 +25,8 @@ export type IObject = OrderFactory.ISimpleOrder & {
  */
 export type IResult = IOwnershipInfo<IGood>[];
 export interface IPotentialActions {
-    /**
-     * 通貨転送アクション
-     */
-    moneyTransfer?: IMoneyTransferActionAttributes[];
-    /**
-     * サービス登録アクション
-     */
-    registerService?: IRegisterServiceAttributes[];
+    // moneyTransfer?: IMoneyTransferActionAttributes[]; // 通貨転送アクション廃止(2024-01-29~)
+    // registerService?: IRegisterServiceAttributes[]; // サービス登録アクション廃止(2024-01-30~)
     /**
      * Eメール送信アクション
      */