+++ src/task/onAssetTransactionStatusChanged.ts
@@ -25,7 +25,18 @@ export interface IObjectAsReserveTransaction {
     transactionNumber: string;
     status: TransactionStatusType.Confirmed;
 }
-export type IObject = IObjectAsPayTransaction | IObjectAsReserveTransaction;
+export interface IObjectAsMoneyTransferTransaction {
+    typeOf: AssetTransactionType.MoneyTransfer;
+    transactionNumber: string;
+    status: TransactionStatusType.Confirmed;
+}
+export interface IObjectAsRegisterServiceTransaction {
+    typeOf: AssetTransactionType.RegisterService;
+    transactionNumber: string;
+    status: TransactionStatusType.Confirmed;
+}
+export type IObject =
+    IObjectAsPayTransaction | IObjectAsReserveTransaction | IObjectAsMoneyTransferTransaction | IObjectAsRegisterServiceTransaction;
 export interface IData {
     project: Pick<IProject, 'id' | 'typeOf'>;
     object: IObject;